+++ src/ng/services/dom/column.js
@@ -1,6 +1,6 @@
 import Cell from './cell';
-import ElementCore from './elementCore';
-import Element from './element';
+import {Element, ElementCore} from './element';
+
 
 class ColumnCore extends ElementCore {
 	constructor() {
