+++ src/ng/services/dom/element.js
@@ -1,6 +1,38 @@
-import ElementCore from './elementCore';
+export class ElementCore {
+	constructor() {
+	}
+
+	get model() {
+		return null;
+	}
+
+	rect() {
+		return {
+			top: 0,
+			left: 0,
+			right: 0,
+			bottom: 0,
+			width: 0,
+			height: 0
+		};
+	}
+
+	addClass() {
+	}
+
+	removeClass() {
+	}
+
+	get width() {
+		return 0;
+	}
+
+	get height() {
+		return 0;
+	}
+}
 const empty = new ElementCore();
-export default class Element {
+export class Element {
 	constructor(element) {
 		this.element = element;
 	}
