--- src/ng/services/dom/elementCore.js
@@ -1,33 +1,0 @@
-export default class ElementCore {
-	constructor() {
-	}
-
-	get model() {
-		return null;
-	}
-
-	rect() {
-		return {
-			top: 0,
-			left: 0,
-			right: 0,
-			bottom: 0,
-			width: 0,
-			height: 0
-		};
-	}
-
-	addClass() {
-	}
-
-	removeClass() {
-	}
-
-	get width() {
-		return 0;
-	}
-
-	get height() {
-		return 0;
-	}
-}
